+++ console_gpt/assistant.py
@@ -35,7 +35,9 @@ def assistant(console, data) -> None:
         elif user_input.lower() in ["flush", "new"]:
             break
         # TODO implement dedicated command handler for assistants
-        handled_user_input = command_handler(data.model["model_title"], data.model["model_name"], user_input, "", False, [])
+        handled_user_input = command_handler(
+            data.model["model_title"], data.model["model_name"], user_input, "", False, []
+        )
         match handled_user_input:
             case "continue" | None:
                 continue