+++ console_gpt/chat.py
@@ -77,7 +77,9 @@ def chat(console, data, managed_user_prompt) -> None:
             if not user_input:  # Used to catch SIGINT
                 save_chat(conversation, ask=True)
             # Command Handler
-            handled_user_input = command_handler(model_title, model_name, user_input["content"], conversation, cached, tools)
+            handled_user_input = command_handler(
+                model_title, model_name, user_input["content"], conversation, cached, tools
+            )
             match handled_user_input:
                 case ("continue", new_tools):
                     tools = new_tools