+++ CHANGELOG.md
@@ -2,6 +2,9 @@
 - [6828df1] [Feature] Update supported models list
 - [7327f53] [Feature] To think or not to think (Sonnet 3.7)
 
+> Improvements:
+- [71de82d] [Improvement] Optimize tools command
+
 > Bug Fixes:
 - [b077a4c] [BUGFIX] Resolved Pillow issue with Python 3.13.x (Fixes: #24)